# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sui-Trace API
  version: 0.1.0
  description: |
    REST surface for the Sui-Trace developer-first explorer.
    A GraphQL endpoint with the equivalent surface is available at `/graphql`,
    and a real-time WebSocket stream at `/ws`.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/package/{id}/source:
    get:
      summary: List decompiled / disassembled module sources for a package
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: per-module summaries (no body) }
    post:
      summary: Push a decompiled module source (external decompiler ingestion)
      description: |
        Authenticated via either a Bearer JWT with role=admin|indexer or the
        configured ingestion shared secret in `X-Trace-Ingest-Key`.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - ingestKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [module_name, source]
              properties:
                module_name: { type: string }
                format:
                  type: string
                  enum: [move-disasm, move-source, pseudo]
                  default: move-disasm
                source: { type: string }
                decompiler: { type: string, default: unknown }
                decompiler_version: { type: string, nullable: true }
                bytecode_hash: { type: string, nullable: true }
      responses:
        "200": { description: stored }
        "400": { description: empty module_name or source }
        "401": { description: missing credential }
        "403": { description: role not allowed }
...
  /api/v1/watch/protocols/{id}:
    get:
      summary: Protocol detail (TVL + code events + activity)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: detail payload }
        "404": { description: not found }
    put:
      summary: Update a protocol (partial)
      security:
        - bearerAuth: []
        - ingestKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "200": { description: updated protocol }
        "404": { description: not found }
    delete:
      summary: Remove a protocol (also drops `protocol_code_events`)
      security:
        - bearerAuth: []
        - ingestKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: removed }
        "404": { description: not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ingestKey:
      type: apiKey
      in: header
      name: X-Trace-Ingest-Key
      description: Shared secret for headless ingestion endpoints (decompiler, etc.).